refactor: Move Covec and FrameTime to other files.
